requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
